Update autocomplete with label option
The label option makes it possible to update an extra observable in your viewmodel.
This allows flattened viewmodels to update key and label of a referenced object.

e.g. I had a viewmodel with personId and personName and wanted both to update in the viewmodel. passing personId as value and personName as label solved this issue.