Update knockout-jqAutocomplete.spec.js
If the test passes with 2000, then shouldn't the test read "override local options with global"? I would think that it would override global (instance.options) with local (passed into init), but at least this fixes the test wording. Or am I not understanding which is local and which is global?